Separate arch specific code to directories
