+++ CMakeLists.txt
@@ -78,6 +78,17 @@ include(${PROJECT_SOURCE_DIR}/cmake/dependencies.cmake)
 add_library(umd_common_directories INTERFACE)
 target_include_directories(umd_common_directories INTERFACE ${PROJECT_SOURCE_DIR} ${PROJECT_SOURCE_DIR}/device ${PROJECT_SOURCE_DIR}/third_party/fmt/include)
 
+if(CMAKE_SYSTEM_PROCESSOR MATCHES "x86_64" OR CMAKE_SYSTEM_PROCESSOR MATCHES "i386")
+    message(STATUS "Building for host arch - x86")
+    target_include_directories(umd_common_directories INTERFACE ${PROJECT_SOURCE_DIR}/device/arch/x86)
+elseif(CMAKE_SYSTEM_PROCESSOR MATCHES "arm" OR CMAKE_SYSTEM_PROCESSOR MATCHES "aarch64")
+    message(STATUS "Building UMD for host arch - ARM")
+    target_include_directories(umd_common_directories INTERFACE ${PROJECT_SOURCE_DIR}/device/arch/aarch64)
+elseif(CMAKE_SYSTEM_PROCESSOR MATCHES "riscv")
+    message(STATUS "Building UMD for host arch - riscv")
+    target_include_directories(umd_common_directories INTERFACE ${PROJECT_SOURCE_DIR}/device/arch/riscv)
+endif()
+
 if(NOT DEFINED ENV{ARCH_NAME})
     message(FATAL_ERROR "Please set ARCH_NAME to grayskull, wormhole_b0, or blackhole")
 elseif($ENV{ARCH_NAME} STREQUAL "grayskull")
