+++ device/device_api_metal.h
@@ -6,4 +6,4 @@
 
 #pragma once
 #include "device/tt_device.h"
-#include "device/driver_atomics.h"
+#include "driver_atomics.h"
