+++ device/simulation/deprecated/tt_versim_device.cpp
@@ -5,7 +5,7 @@
 
 
 #include "tt_device.h"
-#include "device/driver_atomics.h"
+#include "driver_atomics.h"
 #include "common/logger.hpp"
 #include <iostream>
 #include <fstream>
